Revert hero search form back inside the hero section
Restores the main-branch pattern: the search form is a direct child of
.hero__content (after .hero__intro), not a standalone section below the
hero. Form padding, field spacing, and submit-button sizing are restored
to main's compact values. Removes HeroSearchSection.jsx.
