+++ src/styles/homepage/mobile.css
@@ -44,23 +44,20 @@ html, body { max-width: 100%; }
 /* ---- Hero: tame display sizes & padding ---- */
 @media (max-width: 720px) {
   .hero--photo {
-    min-height: 100svh;
-    align-items: center;
-    /* Nav offset on top so content clears the fixed nav; symmetric extra
-       below so the visible area between nav and viewport bottom is the
-       container that vertically centers the content. */
-    padding:
-      calc(var(--nav-height) + max(.5rem, env(safe-area-inset-top, 0px)))
-      clamp(1.25rem, 5vw, 6rem)
-      calc(.5rem + env(safe-area-inset-bottom, 0px));
+    min-height: auto;
+    align-items: flex-start;
+    padding: 0 clamp(1.25rem, 5vw, 6rem) clamp(2.25rem, 5vh, 3.5rem);
   }
   .hero__content { transform: none; }
   .hero__intro {
+    min-height: 100svh;
     width: 100%;
     display: grid;
     justify-items: center;
     align-content: center;
     gap: clamp(.6rem, 1.05vh, 1rem);
+    padding-top: calc(var(--nav-height) + max(1rem, env(safe-area-inset-top, 0px)));
+    padding-bottom: calc(1.25rem + env(safe-area-inset-bottom, 0px));
   }
   .hero__eyebrow { font-size: 10px; padding: .4rem .8rem; margin-bottom: 0; letter-spacing: .18em; }
   .hero__title { font-size: clamp(2.75rem, 13vw, 4.25rem); line-height: 1.02; }
@@ -88,6 +85,8 @@ html, body { max-width: 100%; }
 
   .hero__intro {
     gap: .48rem;
+    padding-top: calc(var(--nav-height) + max(1rem, env(safe-area-inset-top, 0px)));
+    padding-bottom: calc(1rem + env(safe-area-inset-bottom, 0px));
   }
 
   .hero__eyebrow {
@@ -171,14 +170,14 @@ html, body { max-width: 100%; }
   }
 
   .hero__search {
-    margin-top: clamp(1.5rem, 5vh, 2.5rem);
+    margin-top: 0;
   }
 }
 
 @media (max-width: 480px) and (max-height: 700px) {
-  .hero--photo {
-    padding-top: calc(var(--nav-height) + .35rem);
-    padding-bottom: .35rem;
+  .hero__intro {
+    padding-top: calc(var(--nav-height) + .75rem);
+    padding-bottom: .75rem;
   }
 }
 
@@ -194,6 +193,8 @@ html, body { max-width: 100%; }
 
   .hero__intro {
     gap: .34rem;
+    padding-top: calc(var(--nav-height) + .6rem);
+    padding-bottom: .6rem;
   }
 
   .hero__eyebrow {
@@ -232,52 +233,17 @@ html, body { max-width: 100%; }
   }
 
   .hero__search {
-    margin-top: clamp(1.5rem, 5vh, 2.5rem);
+    margin-top: 0;
   }
 }
 
 /* Hero search — cleaner stacked touch targets */
 @media (max-width: 780px) {
-  .hero__search {
-    margin-top: clamp(2rem, 6vh, 3.5rem);
-    padding: .5rem;
-    border-radius: 18px;
-    gap: 4px;
-  }
-  .hero__search-field {
-    padding: .7rem .95rem;
-    border-right: none;
-    border-bottom: 1px solid color-mix(in srgb, var(--text) 12%, transparent);
-  }
-  .hero__search-field:last-of-type { border-bottom: none; }
-  .hero__search-field label {
-    font-size: 10px;
-    letter-spacing: .18em;
-    color: color-mix(in srgb, var(--text) 65%, transparent);
-  }
-  .hero__search-field select,
-  .hero__search-field input {
-    font-size: 16px; /* prevents iOS zoom */
-    color: var(--text);
-  }
-  .hero__search-submit {
-    margin-top: 6px;
-    padding: 1rem;
-    justify-content: center;
-    min-height: 52px;
-    border-radius: 14px;
-    font-size: 15px;
-    font-weight: 700;
-    letter-spacing: .02em;
-    /* Solid coral so the button reads clearly on the translucent dark form */
-    background-image: none;
-    background-color: var(--dp-accent);
-    box-shadow: 0 12px 28px -16px rgba(255, 111, 97, .65);
-  }
-  .hero__search-submit:hover {
-    background-image: none;
-    background-color: color-mix(in srgb, var(--dp-accent) 90%, #000);
-  }
+  .hero__search { margin-top: 1.25rem; padding: .35rem; border-radius: 14px; }
+  .hero__search-field { padding: .65rem .9rem; }
+  .hero__search-field label { font-size: 9px; }
+  .hero__search-field select, .hero__search-field input { font-size: 16px; /* prevents iOS zoom */ }
+  .hero__search-submit { padding: .95rem; justify-content: center; min-height: 48px; border-radius: 10px; }
 }
 
 @supports (content-visibility: auto) {
@@ -286,31 +252,6 @@ html, body { max-width: 100%; }
       content-visibility: auto;
       contain-intrinsic-size: 300px;
     }
-  }
-}
-
-@media (max-width: 480px) {
-  .hero__search {
-    width: 100%;
-    margin-top: .15rem;
-    padding: .45rem;
-    border-radius: 16px;
-  }
-
-  .hero__search-field {
-    padding: .6rem .85rem;
-  }
-
-  .hero__search-submit {
-    min-height: 48px;
-    padding: .85rem .9rem;
-    border-radius: 12px;
-  }
-}
-
-@media (max-width: 480px) and (max-height: 700px) {
-  .hero__search-field {
-    padding-block: .46rem;
   }
 }
