+++ src/styles/homepage/nav-stand.css
@@ -264,21 +264,6 @@
 .hero__scroll-line::after { content: ""; position: absolute; left: 0; right: 0; top: -20px; height: 20px; background: #fff; animation: scrollDot 2.2s ease-in-out infinite; }
 @keyframes scrollDot { 0% { top: -20px; } 100% { top: 48px; } }
 
-/* Standalone search band — matches the next section's surface so it reads
-   as one continuous band with the excursions block below. */
-.hero-search-section {
-  position: relative;
-  margin: 0;
-  padding: clamp(1.5rem, 4vh, 3rem) clamp(1rem, 5vw, 3rem) clamp(2rem, 5vh, 3.5rem);
-  background: var(--surface);
-}
-.hero-search-section .hero__search {
-  width: var(--section-contained-width, min(95%, 1240px));
-  max-width: 880px;
-  margin: 0 auto;
-  animation-delay: .25s;
-}
-
 /* Quick search overlay */
 .hero__search {
   position: relative; z-index: 3;
